Integrated unit test framework and added travis-Ci support
With this commit, we have the ability to write unit tests for the
Autopilot and run them via a command line option. Along with this, the
travis-ci config file has been updated to instruct travis to run all
unit tests on windows, linux, and macos. It has also been updated to
attempt to build the code using the embedded compiler on macos and
windows (previously travis only made sure things built on linux).

Quick breakdown of why each file was added, deleted, or changed:

The gitignore file now ignores the new testBuild directories that are
created when unit testing is taking place.

The travis config file is updated to implement the features mentioned in the
first paragraph.

The CmakeLists file is updated to add the ability to perform a unit test
build.

The build.bash file is updated to allow the new -t option, which
instructs cmake to build the unit tests, rather than build for the
hardware.

The Test directory is added. Inside it, we have the Src and Inc
subdirectories, meant to contain mock class definitions and unit tests,
respectively. Inside it, there is also the vendor subdirectory, which
contains the lib subdirectory (meant to contain unit testing libraries)
and the FindCmakeModules subdirectory, meant to contain Cmake scripts
needed to find unit testing components.

The 2 Test_*_Skeleton files added to the Test/Src directory are meant to
serve as examples for writing unit tests and also allow things to
compile during the unit test build (since the Altimeter and Imu modules
already had partial implementations that attempted to talk to hardware,
they needed to be stubbed out).

The testMain.cpp file serves as a main function for the executable
produced when running unit tests. It should not be messed with.

The fff.h file is the complete fake functions framework, which provides
us with the ability to mock C stuff (while googletest's gmock gives us
the ability to mock C++ stuff). It is being linked in to the unit test
builds.

The FindGMock.cmake file is a script that enables Cmake to locate the
Gmock files on any OS.

Lastly, the /Safety/Common folder is deleted. It was never meant to exist. It
was a symbolic link created by Cmake to cope with a bug that has already been
fixed. Since the build system is being updated in this commit, it is a good
opportunity to get rid of it.